requirements/app.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/app.in
#
alembic==1.4.2            # via flask-migrate
certifi==2020.4.5.2       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask
flask-caching==1.9.0      # via -r requirements/app.in
flask-cors==3.0.8         # via -r requirements/app.in
flask-migrate==2.5.3      # via -r requirements/app.in
flask-sqlalchemy==2.4.3   # via -r requirements/app.in, flask-migrate
flask==1.1.2              # via -r requirements/app.in, flask-caching, flask-cors, flask-migrate, flask-sqlalchemy
idna==2.9                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via jinja2, mako
marshmallow==3.6.1        # via -r requirements/app.in
psycopg2-binary==2.8.5    # via -r requirements/app.in
pyjwt==1.7.1              # via -r requirements/app.in
python-dateutil==2.8.1    # via alembic
python-editor==1.0.4      # via alembic
requests==2.23.0          # via -r requirements/app.in
simplejson==3.17.0        # via -r requirements/app.in
six==1.15.0               # via flask-cors, python-dateutil
sqlalchemy==1.3.17        # via alembic, flask-sqlalchemy
urllib3==1.25.9           # via requests
werkzeug==1.0.1           # via -r requirements/app.in, flask
